Add files via upload
Task 1:- Prepare a complete data analysis report on the given data.

Task 2:- Fix a period for prediction of confirmed cases/deaths. Create a predictive model to forecast the Covid19 cases based on past cases for a specific country or region.

Task3:- Make suggestions to the government health department of the country/region for preparation based on your predictions.